refactor: add a specific state for the create-update actions
- Create a State file for Create Update flow
- Remove Create Update State from Customer State
