+++ spa/src/modules/crm/customers.state.ts
@@ -1,10 +1,8 @@
 import { Customer } from './customer';
 import { customersMock } from './customers.mock';
-import { CreateCustomerStepsModel } from './customers.types';
 
 enum ActionTypes {
   FETCH_CUSTOMERS = '@challenge/crm/customers/FETCH_CUSTOMERS',
-  SAVE_CREATE_CUSTOMER_STEP = '@challenge/crm/customers/SAVE_CREATE_CUSTOMER_STEP',
 }
 
 export interface CustomersAction {
@@ -20,9 +18,6 @@ export default function reducer(
     case ActionTypes.FETCH_CUSTOMERS:
       return action.payload;
 
-    case ActionTypes.SAVE_CREATE_CUSTOMER_STEP:
-      return action.payload;
-
     default:
       return state;
   }
@@ -33,12 +28,5 @@ export const fetchCustomers = () => {
   return {
     type: ActionTypes.FETCH_CUSTOMERS,
     payload: customersMock,
-  };
-};
-
-export const saveCreateCustomerStep = (model: CreateCustomerStepsModel) => {
-  return {
-    type: ActionTypes.SAVE_CREATE_CUSTOMER_STEP,
-    payload: model,
   };
 };
